Break dependency cycle between Batfish and Driver
